% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{actionButton}
\alias{actionButton}
\title{Button input}
\usage{
actionButton(inputId, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{...}{Arguments to be passed to the underlying actionButton function.}
}
\description{
Button input
}
