% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{TextInput}
\alias{TextInput}
\title{TextInput input}
\usage{
TextInput(inputId, ..., value = default_value)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{...}{Arguments to be passed to the underlying textInput function.}

\item{value}{The initial value of the input.}
}
\description{
TextInput input
}
