% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_accordion}
\alias{trem_accordion}
\title{Accordion component}
\usage{
trem_accordion(..., defaultOpen = FALSE)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}

\item{defaultOpen}{Whether the accordion should be open by default.}
}
\description{
Accordion component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-layout-accordion--docs}{Accordion documentation}
}
\seealso{
Other layout: 
\code{\link{trem_card}()},
\code{\link{trem_dialog}()},
\code{\link{trem_divider}()},
\code{\link{trem_flex}()},
\code{\link{trem_grid}()}

Other accordion: 
\code{\link{trem_accordion_body}()},
\code{\link{trem_accordion_header}()},
\code{\link{trem_accordion_list}()}
}
\concept{accordion}
\concept{layout}
