% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_accordion_body}
\alias{trem_accordion_body}
\title{Accordion component}
\usage{
trem_accordion_body(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Accordion component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-layout-accordion--docs}{Accordion documentation}
}
\seealso{
Other accordion: 
\code{\link{trem_accordion}()},
\code{\link{trem_accordion_header}()},
\code{\link{trem_accordion_list}()}
}
\concept{accordion}
