% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_accordion_list}
\alias{trem_accordion_list}
\title{AccordionList component}
\usage{
trem_accordion_list(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
AccordionList component
}
\references{
\href{https://storybook.tremor.so/?path=/story/ui-layout-accordionlist--with-less-than-three-children}{AccordionList documentation}
}
\seealso{
Other accordion: 
\code{\link{trem_accordion}()},
\code{\link{trem_accordion_body}()},
\code{\link{trem_accordion_header}()}
}
\concept{accordion}
