% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{trem_badge}
\alias{trem_badge}
\title{Badge component}
\usage{
trem_badge(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Badge component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-icon-badge--docs}{Badge documentation}
}
\seealso{
Other icon: 
\code{\link{trem_badge_delta}()},
\code{\link{trem_icon}()}
}
\concept{icon}
