% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{trem_badge_delta}
\alias{trem_badge_delta}
\title{BadgeDelta component}
\usage{
trem_badge_delta(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
BadgeDelta component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-icon-badgedelta--docs}{BadgeDelta documentation}
}
\seealso{
Other icon: 
\code{\link{trem_badge}()},
\code{\link{trem_icon}()}
}
\concept{icon}
