% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{trem_button}
\alias{trem_button}
\title{Button input}
\usage{
trem_button(inputId, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{...}{Arguments to be passed to the actionButton.}
}
\description{
Button input
}
