% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{trem_callout}
\alias{trem_callout}
\title{Callout component}
\usage{
trem_callout(title, ...)
}
\arguments{
\item{title}{Title of the callout}

\item{...}{Additional arguments to be passed to the component.}
}
\description{
Callout component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-text-callout--docs}{Callout documentation}
}
\seealso{
Other text: 
\code{\link{trem_legend}()},
\code{\link{trem_metric}()},
\code{\link{trem_subtitle}()},
\code{\link{trem_text}()},
\code{\link{trem_title}()}
}
\concept{text}
