% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_card}
\alias{trem_card}
\title{Card component}
\usage{
trem_card(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Card component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-layout-card--docs}{Card documentation}
}
\seealso{
Other layout: 
\code{\link{trem_accordion}()},
\code{\link{trem_divider}()},
\code{\link{trem_flex}()},
\code{\link{trem_grid}()}
}
\concept{layout}
