% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_dialog}
\alias{trem_dialog}
\title{Dialog component}
\usage{
trem_dialog(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Dialog component
}
\references{
\href{https://storybook.tremor.so/?path=/story/ui-layout-dialog--default}{Dialog documentation}
}
\seealso{
Other layout: 
\code{\link{trem_accordion}()},
\code{\link{trem_card}()},
\code{\link{trem_divider}()},
\code{\link{trem_flex}()},
\code{\link{trem_grid}()}

Other dialog: 
\code{\link{trem_dialog_panel}()}
}
\concept{dialog}
\concept{layout}
