% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_dialog_panel}
\alias{trem_dialog_panel}
\title{#' DialogPanel component}
\usage{
trem_dialog_panel(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
#' DialogPanel component
}
\references{
\href{https://storybook.tremor.so/?path=/story/ui-layout-dialog--default}{DialogPanel documentation}
}
\seealso{
Other dialog: 
\code{\link{trem_dialog}()}
}
\concept{dialog}
