% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_divider}
\alias{trem_divider}
\title{Divider component}
\usage{
trem_divider(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Divider component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-layout-divider--docs}{Divider documentation}
}
\seealso{
Other layout: 
\code{\link{trem_accordion}()},
\code{\link{trem_card}()},
\code{\link{trem_dialog}()},
\code{\link{trem_flex}()},
\code{\link{trem_grid}()}
}
\concept{layout}
