% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_flex}
\alias{trem_flex}
\title{Flex component}
\usage{
trem_flex(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Flex component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-layout-flex--docs}{Flex documentation}
}
\seealso{
Other layout: 
\code{\link{trem_accordion}()},
\code{\link{trem_card}()},
\code{\link{trem_divider}()},
\code{\link{trem_grid}()}
}
\concept{layout}
