% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{trem_grid}
\alias{trem_grid}
\title{Grid component}
\usage{
trem_grid(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Grid component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-layout-grid--docs}{Grid documentation}
}
\seealso{
Other layout: 
\code{\link{trem_accordion}()},
\code{\link{trem_card}()},
\code{\link{trem_divider}()},
\code{\link{trem_flex}()}
}
\concept{layout}
