% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{trem_icon}
\alias{trem_icon}
\title{Icon component}
\usage{
trem_icon(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Icon component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-icon-badgedelta--docs}{Icon documentation}
}
\seealso{
Other icon: 
\code{\link{trem_badge}()},
\code{\link{trem_badge_delta}()}
}
\concept{icon}
