% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{trem_legend}
\alias{trem_legend}
\title{Legend component}
\usage{
trem_legend(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Legend component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-text-legend--docs}{Legend documentation}
}
\seealso{
Other text: 
\code{\link{trem_callout}()},
\code{\link{trem_metric}()},
\code{\link{trem_subtitle}()},
\code{\link{trem_text}()},
\code{\link{trem_title}()}
}
\concept{text}
