% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{trem_metric}
\alias{trem_metric}
\title{Metric component}
\usage{
trem_metric(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Metric component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-text-metric--docs}{Metric documentation}
}
\seealso{
Other text: 
\code{\link{trem_callout}()},
\code{\link{trem_legend}()},
\code{\link{trem_subtitle}()},
\code{\link{trem_text}()},
\code{\link{trem_title}()}
}
\concept{text}
