% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{trem_subtitle}
\alias{trem_subtitle}
\title{Subtitle component}
\usage{
trem_subtitle(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Subtitle component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-text-subtitle--docs}{Subtitle documentation}
}
\seealso{
Other text: 
\code{\link{trem_callout}()},
\code{\link{trem_legend}()},
\code{\link{trem_metric}()},
\code{\link{trem_text}()},
\code{\link{trem_title}()}
}
\concept{text}
