% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{trem_title}
\alias{trem_title}
\title{Title component}
\usage{
trem_title(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the component.}
}
\description{
Title component
}
\references{
\href{https://storybook.tremor.so/?path=/docs/ui-text-title--docs}{Title documentation}
}
\seealso{
Other text: 
\code{\link{trem_callout}()},
\code{\link{trem_legend}()},
\code{\link{trem_metric}()},
\code{\link{trem_subtitle}()},
\code{\link{trem_text}()}
}
\concept{text}
