% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tremousser-package.R
\docType{package}
\name{tremousser-package}
\alias{tremousser}
\alias{tremousser-package}
\title{tremousser: Tremor Library For Shiny}
\description{
A kit of tools to build charts and dashboards in Shiny with the Tremor library.
}
\author{
\strong{Maintainer}: Arthur Bréant \email{arthur@thinkr.fr} (\href{https://orcid.org/0000-0003-1668-0963}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
